+++ ui/packages/shared/functions/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.16.58](https://github.com/parca-dev/parca/compare/@parca/functions@0.16.57...@parca/functions@0.16.58) (2023-01-19)
+
+**Note:** Version bump only for package @parca/functions
+
+## 0.16.57 (2023-01-17)
+
+**Note:** Version bump only for package @parca/functions
+
 ## 0.16.56 (2023-01-11)
 
 **Note:** Version bump only for package @parca/functions